Add BuildingType registry and mod XMLs
Introduce a C++ BuildingType registry to load XML-based building graphic rules from Mods (Mods/Augustus and Mods/Vespasian). Adds src/building/building_type_registry.{cpp,h} and new BuildingType XML files, and integrates them into the solution/project (solution name/TargetName changed to Vespasian, project files updated to compile and include the new sources and copy mod XMLs to output). Convert src/building/figure.c to C++ (figure.cpp) and replace numerous per-building graphic setters with building_type_registry_apply_graphic to centralize graphic/upgrade logic. Add API to reset runtime instances and set/get active mod name; call reset in several game lifecycle points. Also update save file extension handling to .savf, add save/mod mismatch UI helper, start persisting save mod metadata (file_io changes), and include a few minor fixes/cleanups (type casts, integer math, placeholder function).

+++ src/graphics/lang_text.c
@@ -530,5 +530,5 @@ int lang_text_concatenate_sequence(const lang_fragment *seq, int count, uint8_t 
     }
 
     *cursor = 0; // Null terminate
-    return cursor - dst; // Return length of concatenated string
+    return (int) (cursor - dst); // Return length of concatenated string
 }
